tests/baseline/protocol-1.2/inspector/sources/debugger-async/async-callstack-in-console-expected.txt

CONSOLE MESSAGE: line 11: console.clear
CONSOLE MESSAGE: line 17: console.trace
CONSOLE ERROR: line 24: Uncaught Error: foo
CONSOLE MESSAGE: line 29: console.trace
CONSOLE ERROR: line 37: Failed to execute 'postMessage' on 'DOMWindow': The target origin provided ('http://www.example.com') does not match the recipient window's origin ('').
Tests asynchronous call stacks printed in console.


Set timer for test function.
async-callstack-in-console.html:11 Console was cleared
async-callstack-in-console.html:17 console.trace

timeout1 @ async-callstack-in-console.html:17

setTimeout (async)

testFunction @ async-callstack-in-console.html:12
async-callstack-in-console.html:24 Uncaught Error: foo
    at timeout2 (async-callstack-in-console.html:24)

timeout2 @ async-callstack-in-console.html:24

setTimeout (async)

timeout1 @ async-callstack-in-console.html:18

setTimeout (async)

testFunction @ async-callstack-in-console.html:12
async-callstack-in-console.html:29 console.trace

timeout3 @ async-callstack-in-console.html:29

setTimeout (async)

timeout2 @ async-callstack-in-console.html:23

setTimeout (async)

timeout1 @ async-callstack-in-console.html:18

setTimeout (async)

testFunction @ async-callstack-in-console.html:12
async-callstack-in-console.html:37 Failed to execute 'postMessage' on 'DOMWindow': The target origin provided ('http://www.example.com') does not match the recipient window's origin ('file://').

tryPostMessage @ async-callstack-in-console.html:37

timeout3 @ async-callstack-in-console.html:31

setTimeout (async)

timeout2 @ async-callstack-in-console.html:23

setTimeout (async)

timeout1 @ async-callstack-in-console.html:18

setTimeout (async)

testFunction @ async-callstack-in-console.html:12
